refactor(okta): extract DTO classes to separate files in Okta/ subfolder
Extract IdpTokenRestRequest, IdpTokenRequest, IdpTokenResponse, and
SamlRestRequest classes from OktaAuthenticator.cs to individual files
in the new Okta/ subfolder. This improves code organization and follows
the single responsibility principle.
